Move BookList to container
